cambios en base de datos y pequeños detalles

+++ Agraria.Modelo/Entidades/HVentas_Detalle.cs
@@ -13,7 +13,7 @@ namespace Agraria.Modelo.Entidades
         public string? Cod_Art { get; set; }
         public string? Descr { get; set; }
         public decimal P_Unit { get; set; }
-        public int Cant { get; set; }
+        public decimal Cant { get; set; }
         public decimal P_X_Cant { get; set; }
 
         public HVentasDetalle() { }
